internal representation of distances are now floats, and networkLatency returns a float improved destroy() pattern   PastryEndpoint checks if its application is Destructable, and if so, calls destroy.   made MessageDispatch.destroy() get called, (was a regression bug)   node network coordinates can be reused   leafset and routing table.destroy() now are called and release handles to dead nodes, by calling deleteObserver()   cleared observers/references in destroy() method of Join/Leafset protocols many changes to RoutingTableTest
git-svn-id: https://svn.mpi-sws.org/DS/freepastry/trunk@3418 c21dacaf-8602-0410-aacd-8bc41df2cbe7

+++ pastry/src/rice/pastry/direct/BasicNetworkSimulator.java
@@ -323,7 +323,7 @@ public abstract class BasicNetworkSimulator implements NetworkSimulator {
    * @param b the second NodeId
    * @return the proximity between the two input NodeIds
    */
-  public double networkDelay(DirectNodeHandle a, DirectNodeHandle b) {
+  public float networkDelay(DirectNodeHandle a, DirectNodeHandle b) {
     NodeRecord nra = a.getRemote().record;
     NodeRecord nrb = b.getRemote().record;
 
@@ -334,7 +334,7 @@ public abstract class BasicNetworkSimulator implements NetworkSimulator {
     return nra.networkDelay(nrb);
   }
 
-  public int proximity(DirectNodeHandle a, DirectNodeHandle b) {
+  public float proximity(DirectNodeHandle a, DirectNodeHandle b) {
     NodeRecord nra = a.getRemote().record;
     NodeRecord nrb = b.getRemote().record;
 
@@ -357,12 +357,12 @@ public abstract class BasicNetworkSimulator implements NetworkSimulator {
   public DirectNodeHandle getClosest(DirectNodeHandle nh) {
     Iterator it = nodes.iterator();
     DirectNodeHandle bestHandle = null;
-    int bestProx = Integer.MAX_VALUE;
+    float bestProx = Float.MAX_VALUE;
     Id theId;
 
     while (it.hasNext()) {
       DirectPastryNode theNode = (DirectPastryNode) it.next();
-      int theProx = theNode.record.proximity(nh.getRemote().record);
+      float theProx = theNode.record.proximity(nh.getRemote().record);
       theId = theNode.getNodeId();
       if (!theNode.isAlive() || !theNode.isReady()
           || theId.equals(nh.getNodeId())) {
